Improve Playgrounds by moving SwiftUI to new pages
