+++ ParseSwift.playground/Pages/1 - Your first Object.xcplaygroundpage/Contents.swift
@@ -309,5 +309,4 @@ do {
 }*/
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
