+++ ParseSwift.playground/Pages/10 - Cloud Code.xcplaygroundpage/Contents.swift
@@ -113,5 +113,4 @@ score.save(options: [.context(["hello": "world"])]) { result in
 }
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
