+++ ParseSwift.playground/Pages/12 - Roles and Relations.xcplaygroundpage/Contents.swift
@@ -229,5 +229,4 @@ let relation = User.current!.relation
 //: similar to `users` and `roles`.
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
