+++ ParseSwift.playground/Pages/2 - Finding Objects.xcplaygroundpage/Contents.swift
@@ -129,5 +129,4 @@ queryExclude.first { results in
 }
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
