+++ ParseSwift.playground/Pages/3 - User - Sign Up.xcplaygroundpage/Contents.swift
@@ -57,5 +57,4 @@ User.signup(username: "hello", password: "world") { results in
 }
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
