+++ ParseSwift.playground/Pages/4 - User - Continued.xcplaygroundpage/Contents.swift
@@ -213,5 +213,4 @@ User.current?.signup { result in
 }
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
