+++ ParseSwift.playground/Pages/5 - ACL.xcplaygroundpage/Contents.swift
@@ -66,5 +66,4 @@ score.save { result in
 }
 
 PlaygroundPage.current.finishExecution()
-
 //: [Next](@next)
